# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Flash Cell Store API
  version: '1.0'
  description: |
    API do sistema Flash Cell Store — loja virtual, assistência técnica e
    painel administrativo.

    **Autenticação:** a maioria dos endpoints exige um `Authorization: Bearer
    <accessToken>`, obtido em `/auth/login` ou `/auth/register`. Access
    tokens expiram em 15 minutos; use `/auth/refresh` com o `refreshToken`
    para renovar sem precisar logar de novo.

    **Formato de resposta:** toda resposta segue `{ status: 'ok' | 'error',
    data, meta?, message? }`. Listagens paginadas incluem `meta: { total,
    page, perPage, totalPages }`.
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /products:
    get:
      tags: [Catálogo]
      summary: Lista produtos com filtros
      security: []
      parameters:
        - { name: page, in: query, schema: { type: integer } }
        - { name: perPage, in: query, schema: { type: integer } }
        - { name: search, in: query, schema: { type: string } }
        - { name: brandId, in: query, schema: { type: string, format: uuid } }
        - { name: categoryId, in: query, schema: { type: string, format: uuid } }
        - { name: minPrice, in: query, schema: { type: number } }
        - { name: maxPrice, in: query, schema: { type: number } }
        - { name: isFeatured, in: query, schema: { type: boolean } }
      responses:
        '200': { description: Lista de produtos }
    post:
      tags: [Administração]
      summary: Cria produto (admin)
      responses:
        '201': { description: Produto criado }
  /products/slug/{slug}:
    get:
      tags: [Catálogo]
      summary: Detalhe do produto por slug
      security: []
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Produto com imagens e estoque }
        '404': { description: Produto não encontrado }
  /products/{id}:
    get:
      tags: [Catálogo]
      summary: Detalhe do produto por id
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Produto com imagens e estoque }
...
  /stock/movements:
    get:
      tags: [Estoque]
      summary: Histórico de movimentações (staff)
      parameters:
        - { name: productId, in: query, schema: { type: string, format: uuid } }
        - { name: type, in: query, schema: { type: string, enum: [in, out, adjustment, return, loss] } }
      responses:
        '200': { description: Histórico de movimentações }
    post:
      tags: [Estoque]
      summary: Registra uma movimentação de estoque (staff)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, type, quantity]
              properties:
                productId: { type: string, format: uuid }
                type: { type: string, enum: [in, out, adjustment, return, loss] }
                quantity: { type: integer }
                notes: { type: string }
      responses:
        '201': { description: Movimentação registrada, saldo atualizado }
        '409': { description: Estoque insuficiente para o tipo de movimentação }
...
  /orders:
    get:
      tags: [Carrinho e Pedidos]
      summary: Lista pedidos (próprios, ou todos se staff)
      responses:
        '200': { description: Lista de pedidos }
    post:
      tags: [Carrinho e Pedidos]
      summary: Cria pedido a partir do carrinho (checkout)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                addressId: { type: string, format: uuid }
                couponCode: { type: string }
      responses:
        '201': { description: Pedido criado, estoque baixado, carrinho limpo }
        '409': { description: Estoque insuficiente para algum item }
  /orders/{id}/cancel:
    patch:
      tags: [Carrinho e Pedidos]
      summary: Cancela o pedido e reverte o estoque
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Pedido cancelado, estoque revertido }
        '409': { description: 'Pedido não pode mais ser cancelado (já enviado)' }
...
  /warranties/lookup:
    get:
      tags: [Garantias]
      summary: Consulta pública de garantia por IMEI/série ou CPF/CNPJ
      security: []
      parameters:
        - { name: imei, in: query, schema: { type: string } }
        - { name: document, in: query, schema: { type: string } }
      responses:
        '200': { description: Garantias encontradas }
  /warranties/{id}/claim:
    post:
      tags: [Garantias]
      summary: Registra o acionamento de uma garantia (staff)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Acionamento registrado }
        '409': { description: Garantia expirada ou anulada }
...